docos and tidy up
