+++ src/initiative/prompt-for-initiative.ts
@@ -1,0 +1,72 @@
+/**
+ * @file This module shows a dialog when making an initiative check, allowing the user to roll
+ * manually (instead of relying on RNG).
+ */
+
+import { replaceOnce } from '../helpers/replace-once'
+import { promptForRoll } from './dialog'
+
+/**
+ * Augment the options parameter of the ‘roll’ functions:
+ */
+type RollInitiativeOptions = Parameters<Combat['rollInitiative']>[1] & {
+    /**
+     * If set to false, then this roll will not show a dialog and instead use RNG.
+     */
+    dialog?: false
+}
+
+/**
+ * Mix functionality into the @see Combat class.
+ */
+Hooks.on('init', function()
+{
+    CONFIG.Combat.entityClass = class extends CONFIG.Combat.entityClass
+    {
+        rollAll(options?: RollInitiativeOptions)
+        {
+            // Never show the dialog when the ‘roll all’ button is used:
+            options = { ...options, dialog: false }
+            return super.rollAll(options)
+        }
+
+        rollNPC(options?: RollInitiativeOptions)
+        {
+            // Never show the dialog when the ‘roll NPCs’ button is used:
+            options = { ...options, dialog: false }
+            return super.rollNPC(options)
+        }
+
+        async rollInitiative(idsObj: string[] | string, options?: RollInitiativeOptions): Promise<Combat>
+        {
+            // With this option set, just do the default behaviour:
+            if (options?.dialog === false)
+                return super.rollInitiative(idsObj, options)
+
+            // Get the ID from the first parameter:
+            const id = idsObj instanceof Array ? idsObj[0] : idsObj
+
+            // Get the turn data and show the dialog:
+            const turn = this.turns.find(t => t._id == id)!
+            const rollResult = await promptForRoll(turn)
+
+            // If the user clicked “Use RNG instead,” just do the default behaviour:
+            if (rollResult == 'use-rng')
+                return super.rollInitiative(idsObj, options)
+
+            // If the user clicked cancel, do nothing:
+            if (rollResult == 'cancel')
+                return this
+
+            // If the user entered a number, then override the initiative ‘formula’ for this roll:
+            if (typeof rollResult == 'number')
+            {
+                let formula = (game.system.data as any).initiative
+                formula = replaceOnce(formula, '1d20', rollResult + '')
+                return super.rollInitiative(idsObj, { ...options, formula })
+            }
+
+            throw new Error(`Unexpected roll result: ${rollResult}`)
+        }
+    }
+})
